Move `Layers` to dedicated module
